Implement mdx transformer. Fixing AB#6, AB#1
