+++ content/project/tauims/first-post.mdx
@@ -1,0 +1,13 @@
+---
+title: "My First Post"
+date: "2021-07-23"
+slug: "my-first-post"
+---
+
+This is my first blog post! Isn't it *great*?
+
+Some of my **favorite** things are:
+
+* Petting dogs
+* Singing
+* Eating potato-based foods
