+++ content/project/tauims/last-post.mdx
@@ -1,0 +1,7 @@
+---
+title: "Yet Another Post"
+date: "2021-07-25"
+slug: "yet-another-post"
+---
+
+Wow look at all this content. How do they do it?
