+++ content/project/tauims/second-post.mdx
@@ -1,0 +1,7 @@
+---
+title: "Another Post"
+date: "2021-07-24"
+slug: "another-post"
+---
+
+Here's another post! It's even better than the first one!
