+++ package.json
@@ -10,10 +10,12 @@
   "dependencies": {
     "@fontsource/merriweather": "^4.5.14",
     "@fontsource/montserrat": "^4.5.14",
+    "@mdx-js/react": "^2.2.1",
     "gatsby": "^5.4.2",
     "gatsby-plugin-feed": "^5.4.0",
     "gatsby-plugin-image": "^3.4.0",
     "gatsby-plugin-manifest": "^5.4.0",
+    "gatsby-plugin-mdx": "^5.5.0",
     "gatsby-plugin-sharp": "^5.4.0",
     "gatsby-remark-images": "^7.4.0",
     "gatsby-remark-prismjs": "^7.4.0",
