+++ src/pages/project.js
@@ -1,0 +1,58 @@
+import * as React from 'react'
+import Layout from '../components/layout/layout'
+import Seo from '../components/seo'
+import PropTypes from 'prop-types'
+import {graphql} from 'gatsby'
+
+const ProjectPage = ({location, data}) => {
+  return (
+    <Layout location={location}>
+      {data.allMdx.nodes.map(node => (
+        <article key={node.id}>
+          <h2>{node.frontmatter.title}</h2>
+          <p>
+            Posted: {node.frontmatter.date} | Last updated:{' '}
+            {node.parent.modifiedTime}
+          </p>
+          <p>{node.excerpt}</p>
+        </article>
+      ))}
+    </Layout>
+  )
+}
+
+ProjectPage.propTypes = {
+  data: PropTypes.shape({
+    allMdx: PropTypes.shape({
+      nodes: PropTypes.array,
+    }),
+  }),
+  location: PropTypes.shape({
+    pathname: PropTypes.string.isRequired,
+  }).isRequired,
+}
+
+export const query = graphql`
+  query {
+    allMdx(sort: {frontmatter: {date: DESC}}) {
+      nodes {
+        frontmatter {
+          date(formatString: "MMMM D, YYYY")
+          title
+        }
+        parent {
+          ... on File {
+            modifiedTime(formatString: "MMMM D, YYYY")
+          }
+        }
+        id
+        excerpt
+      }
+    }
+  }
+`
+export const Head = () => (
+  <Seo title="Project Posts" description="List of project posts" />
+)
+
+export default ProjectPage
